+++ provider/gpkg/conn_pool.go
@@ -1,3 +1,5 @@
+// +build cgo
+
 package gpkg
 
 import (
