+++ provider/gpkg/conn_pool_internal_test.go
@@ -1,3 +1,5 @@
+// +build cgo
+
 package gpkg
 
 import (
