+++ provider/gpkg/gpkg_test.go
@@ -1,3 +1,5 @@
+// +build cgo
+
 package gpkg_test
 
 import (
